// src/app/Model/loan.ts
export class Loan {
  companyId?: any;
  divisionId?: any;
  departmentId: any;
  loanTypeId:any;
  description: any;
  interestRate:any;
  interestTypeId?: string | any;
  glaccountNumber:any;
  lockedBy?: any;
  lockTs?: any;
  branchCode?: any;
  data:string;
  status:string;
}

// export interface LoanResponse {
//   loans: Loan []
// }